Update jest, svelte-jest, ts-jest. Install svelte2tsx for better it's type defintions. Moves types from component module to ts files.
Reasoning: Unit tests fail when types are imported in a ts file from a component directly. See https://github.com/sveltejs/svelte/issues/5817.
