+++ mathesar_ui/jest.config.cjs
@@ -11,6 +11,7 @@ function getAlias() {
 }
 
 module.exports = {
+  testEnvironment: 'jsdom',
   transform: {
     '^.+\\.svelte$': [
       'svelte-jester',
