+++ mathesar_ui/package.json
@@ -51,7 +51,7 @@
     "eslint-plugin-jest": "^24.7.0",
     "eslint-plugin-promise": "^5.2.0",
     "eslint-plugin-svelte3": "^3.4.0",
-    "jest": "^26.6.3",
+    "jest": "^27.5.1",
     "prettier": "^2.5.0",
     "prettier-plugin-svelte": "^2.5.0",
     "sass": "^1.35.1",
@@ -59,11 +59,12 @@
     "style-loader": "^3.3.1",
     "svelte": "^3.42.4",
     "svelte-check": "^2.2.5",
-    "svelte-jester": "^1.7.0",
+    "svelte-jester": "^2.3.2",
     "svelte-loader": "^3.1.2",
     "svelte-preprocess": "^4.8.0",
+    "svelte2tsx": "^0.5.9",
     "tinro": "^0.6.4",
-    "ts-jest": "^26.5.6",
+    "ts-jest": "^27.1.4",
     "tslib": "^2.3.0",
     "typescript": "^4.4.4",
     "vite": "^2.5.3"
