+++ mathesar_ui/src/component-library/common/base-components/BaseInput.svelte
@@ -1,17 +1,8 @@
-<script context="module" lang="ts">
-  export interface BaseInputProps {
-    id?: string;
-    labelController?: LabelController;
-    disabled?: boolean;
-    focusOnMount?: boolean;
-  }
-</script>
-
 <script lang="ts">
   import { tick, afterUpdate } from 'svelte';
-  import type { LabelController } from '@mathesar-component-library-dir/label/LabelController';
   import { getLabelControllerFromContainingLabel } from '@mathesar-component-library-dir/label/LabelController';
   import { getGloballyUniqueId } from '@mathesar-component-library-dir/common/utils/domUtils';
+  import type { BaseInputProps } from './BaseInputTypes';
 
   type $$Props = BaseInputProps;
