+++ mathesar_ui/src/component-library/common/base-components/BaseInputTypes.ts
@@ -1,0 +1,8 @@
+import type { LabelController } from '@mathesar-component-library-dir/label/LabelController';
+
+export interface BaseInputProps {
+  id?: string;
+  labelController?: LabelController;
+  disabled?: boolean;
+  focusOnMount?: boolean;
+}
