+++ mathesar_ui/src/component-library/number-input/number-formatter/AbstractNumberFormatter.ts
@@ -1,7 +1,7 @@
 import type {
   InputFormatter,
   ParseResult,
-} from '@mathesar-component-library-dir/formatted-input/InputFormatter';
+} from '@mathesar-component-library-dir/formatted-input/FormattedInputTypes';
 import type { DerivedOptions, Options } from './options';
 import { defaultOptions, getDerivedOptions } from './options';
