+++ mathesar_ui/src/component-library/number-input/number-formatter/NumberFormatter.ts
@@ -1,4 +1,4 @@
-import type { ParseResult } from '@mathesar-component-library-dir/formatted-input/InputFormatter';
+import type { ParseResult } from '@mathesar-component-library-dir/formatted-input/FormattedInputTypes';
 import AbstractNumberFormatter from './AbstractNumberFormatter';
 import { makeFormatter } from './formatter';
 import { makeUniversalNumberParser } from './parsers';
