+++ mathesar_ui/src/component-library/select/SelectTypes.ts
@@ -1,0 +1,17 @@
+import type { BaseInputProps } from '@mathesar-component-library-dir/common/base-components/BaseInputTypes';
+import type { LabelGetter } from '@mathesar-component-library-dir/common/utils/formatUtils';
+import type { Appearance } from '@mathesar-component-library-dir/commonTypes';
+import type { ListBoxProps } from '@mathesar-component-library-dir/list-box/ListBoxTypes';
+
+export interface SelectProps<Option> extends BaseInputProps {
+  options: Option[];
+  value?: Option;
+  labelKey?: string;
+  getLabel?: LabelGetter<Option | undefined>;
+  prependBlank?: boolean;
+  contentClass?: string;
+  triggerClass?: string;
+  triggerAppearance?: Appearance;
+  ariaLabel?: string;
+  valuesAreEqual?: ListBoxProps<Option | undefined>['checkEquality'];
+}
