+++ mathesar_ui/src/component-library/text-area/TextArea.svelte
@@ -1,19 +1,6 @@
-<script context="module" lang="ts">
-  import type { BaseInputProps } from '@mathesar-component-library-dir/common/base-components/BaseInput.svelte';
-
-  // eslint-disable-next-line no-undef
-  type InputProps = svelte.JSX.HTMLAttributes<HTMLElementTagNameMap['input']>;
-  type SimplifiedInputProps = Omit<InputProps, 'disabled' | 'id' | 'class'>;
-
-  export interface TextAreaProps extends SimplifiedInputProps, BaseInputProps {
-    value?: string | null;
-    class?: string;
-    element?: HTMLTextAreaElement;
-  }
-</script>
-
 <script lang="ts">
   import BaseInput from '@mathesar-component-library-dir/common/base-components/BaseInput.svelte';
+  import type { TextAreaProps } from '@mathesar-component-library-dir/text-area/TextAreaTypes';
 
   type $$Props = TextAreaProps;
