+++ mathesar_ui/src/component-library/text-area/TextAreaTypes.ts
@@ -1,0 +1,8 @@
+import type { BaseInputProps } from '@mathesar-component-library-dir/common/base-components/BaseInputTypes';
+import type { SimplifiedInputProps } from '@mathesar-component-library-dir/commonTypes';
+
+export interface TextAreaProps extends SimplifiedInputProps, BaseInputProps {
+  value?: string | null;
+  class?: string;
+  element?: HTMLTextAreaElement;
+}
