+++ mathesar_ui/src/component-library/text-input/TextInput.svelte
@@ -1,20 +1,6 @@
-<script context="module" lang="ts">
-  import type { BaseInputProps } from '@mathesar-component-library-dir/common/base-components/BaseInput.svelte';
-
-  // eslint-disable-next-line no-undef
-  type InputProps = svelte.JSX.HTMLAttributes<HTMLElementTagNameMap['input']>;
-  type SimplifiedInputProps = Omit<InputProps, 'disabled' | 'id' | 'class'>;
-
-  export interface TextInputProps extends SimplifiedInputProps, BaseInputProps {
-    value?: string | null;
-    class?: string;
-    element?: HTMLInputElement;
-    hasError?: boolean;
-  }
-</script>
-
 <script lang="ts">
   import BaseInput from '@mathesar-component-library-dir/common/base-components/BaseInput.svelte';
+  import type { TextInputProps } from './TextInputTypes';
 
   type $$Props = TextInputProps;
