+++ mathesar_ui/src/component-library/types.ts
@@ -1,12 +1,5 @@
 export * from './common/utils';
-
-export type Appearance =
-  | 'default'
-  | 'primary'
-  | 'secondary'
-  | 'plain'
-  | 'ghost';
-export type Size = 'small' | 'medium' | 'large';
+export * from './commonTypes';
 
 export type { Tab } from './tabs/TabContainerTypes';
 export type { TreeItem } from './tree/TreeTypes';
@@ -17,7 +10,7 @@ export * from './dynamic-input/types';
 export * from './form-builder/types';
 export * from './cancel-or-proceed-button-pair/CancelOrProceedButtonPairTypes';
 export * from './list-box/ListBoxTypes';
-export type { TextInputProps } from './text-input/TextInput.svelte';
-export type { TextAreaProps } from './text-area/TextArea.svelte';
-export type { StringifiedNumberInputProps } from './number-input/StringifiedNumberInput.svelte';
-export type { SelectProps } from './select/Select.svelte';
+export * from './text-input/TextInputTypes';
+export * from './text-area/TextAreaTypes';
+export * from './select/SelectTypes';
+export * from './number-input/NumberInputTypes';
